refactor: integrate utility, update, and feedback handler to system
