--- apps/electron/src/main/infrastructure/ipc/handlers/utility-handler.ts
@@ -1,10 +1,0 @@
-import { ipcMain } from "electron";
-import { commandExists } from "@/main/utils/env-utils";
-
-export function setupUtilityHandlers(): void {
-  // Check if a command exists in user shell environment
-  ipcMain.handle("command:exists", async (_, command: string) => {
-    const result = await commandExists(command);
-    return result;
-  });
-}
